# message_service/docs/swagger.yaml
basePath: /
definitions:
...
  controllers.CheckContainerStatusResponse:
    properties:
      status:
        type: string
    type: object
...
  controllers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  controllers.GetContainerLogsResponse:
    properties:
      logs:
        description: Изменено с []string на string
        type: string
    type: object
  controllers.GetRolesResponse:
    properties:
      roles:
        description: Изменено с []models.Role на []string
        items:
          type: string
        type: array
    type: object
...
  controllers.ModerateMessagesResponse:
    properties:
      message:
        type: string
    type: object
...
  controllers.UpdatePasswordRequest:
    properties:
      new_password:
        type: string
      username:
        type: string
    required:
    - new_password
    - username
    type: object
  controllers.UpdatePasswordResponse:
    properties:
      message:
        type: string
    type: object
  controllers.UpdateRolesRequest:
    properties:
      new_role:
        type: string
      user_id:
        type: integer
    required:
    - new_role
    - user_id
    type: object
  controllers.UpdateRolesResponse:
    properties:
      message:
        type: string
    type: object
  controllers.UpdateUserResponse:
    properties:
      message:
        type: string
    type: object
  controllers.UploadAttachmentResponse:
    properties:
      attachment_id:
        type: integer
      filename:
        type: string
      filesize:
        type: integer
      filetype:
        type: string
      message:
        type: string
      upload_time:
        type: string
    type: object
...
  models.Role:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.Role'
        description: Связь с моделью Role
      role_id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /admin/logs:
    get:
      consumes:
      - application/json
      description: Этот эндпоинт возвращает логи контейнеров.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.GetContainerLogsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Получение логов контейнеров
      tags:
      - Контейнеры
  /admin/moderate:
    get:
      consumes:
      - application/json
      description: Этот эндпоинт выполняет модерацию сообщений.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.ModerateMessagesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Модерация сообщений
      tags:
      - Модерация
  /admin/roles:
    get:
      consumes:
      - application/json
      description: Этот эндпоинт возвращает список всех ролей пользователей.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.GetRolesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Получение ролей пользователей
      tags:
      - Роли
    post:
      consumes:
      - application/json
      description: Этот эндпоинт обновляет роли пользователей.
      parameters:
      - description: Информация о ролях
        in: body
        name: roles
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateRolesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.UpdateRolesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Обновление ролей пользователей
      tags:
      - Роли
  /admin/status:
    get:
      consumes:
      - application/json
      description: Этот эндпоинт проверяет загруженность контейнеров.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.CheckContainerStatusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Проверка статуса контейнеров
      tags:
      - Контейнеры
  /attachment:
    post:
      consumes:
      - multipart/form-data
      description: Этот эндпоинт загружает файл как вложение и сохраняет его в базе
        данных.
      parameters:
      - description: Файл для загрузки
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.UploadAttachmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Загрузка вложения
      tags:
      - Вложения
  /attachment/{id}:
    get:
      description: Этот эндпоинт скачивает вложение по его ID.
      parameters:
      - description: ID вложения
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Вложение
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Скачивание вложения
      tags:
      - Вложения
...
  /user/update:
    put:
      consumes:
      - application/json
      description: Этот эндпоинт обновляет информацию о пользователе.
      parameters:
      - description: Информация о пользователе
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.UpdateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Обновление данных пользователя
      tags:
      - Пользователи
  /user/update-password:
    put:
      consumes:
      - application/json
      description: Этот эндпоинт меняет пароль пользователя.
      parameters:
      - description: Запрос на смену пароля
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.UpdatePasswordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Смена пароля пользователя
      tags:
      - Пользователи
swagger: "2.0"